requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy>=1.26.0
pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-image>=0.22.0
